Added dtype safeguards to the 'required_model_output_shape()' methods that cast from np.float32 to np.int8.

+++ rllib/models/tf/tf_action_dist.py
@@ -188,6 +188,7 @@ class MultiCategorical(TFActionDistribution):
             np.product(action_space.shape) * (high_ - low_ + 1)
         # MultiDiscrete space.
         else:
+            # nvec is already integer, so no casting needed.
             return np.sum(action_space.nvec)
 
 
@@ -375,7 +376,7 @@ class DiagGaussian(TFActionDistribution):
     def required_model_output_shape(
         action_space: gym.Space, model_config: ModelConfigDict
     ) -> Union[int, np.ndarray]:
-        return np.prod(action_space.shape) * 2
+        return np.prod(action_space.shape, dtype=np.int8) * 2
 
 
 @DeveloperAPI
@@ -474,7 +475,7 @@ class SquashedGaussian(TFActionDistribution):
     def required_model_output_shape(
         action_space: gym.Space, model_config: ModelConfigDict
     ) -> Union[int, np.ndarray]:
-        return np.prod(action_space.shape) * 2
+        return np.prod(action_space.shape, dtype=np.int8) * 2
 
 
 @DeveloperAPI
@@ -530,7 +531,7 @@ class Beta(TFActionDistribution):
     def required_model_output_shape(
         action_space: gym.Space, model_config: ModelConfigDict
     ) -> Union[int, np.ndarray]:
-        return np.prod(action_space.shape) * 2
+        return np.prod(action_space.shape, dtype=np.int8) * 2
 
 
 @DeveloperAPI
@@ -658,7 +659,7 @@ class MultiActionDistribution(TFActionDistribution):
 
     @override(ActionDistribution)
     def required_model_output_shape(self, action_space, model_config):
-        return np.sum(self.input_lens)
+        return np.sum(self.input_lens, dtype=np.int8)
 
 
 @DeveloperAPI
@@ -715,4 +716,4 @@ class Dirichlet(TFActionDistribution):
     def required_model_output_shape(
         action_space: gym.Space, model_config: ModelConfigDict
     ) -> Union[int, np.ndarray]:
-        return np.prod(action_space.shape)
+        return np.prod(action_space.shape, dtype=np.int8)
